refactor(common): CHECKOUT-2951 Convert `geography` module into TypeScript
