+++ src/geography/country-reducer.ts
@@ -1,13 +1,15 @@
-import { combineReducers } from '@bigcommerce/data-store';
+import { combineReducers, Action } from '@bigcommerce/data-store';
+
 import * as actionTypes from './country-action-types';
 
 /**
+ * @todo Convert this file into TypeScript properly
  * @param {CountriesState} state
  * @param {Action} action
  * @return {CountriesState}
  */
-export default function countryReducer(state = {}, action) {
-    const reducer = combineReducers({
+export default function countryReducer(state: any = {}, action: Action): any {
+    const reducer = combineReducers<any>({
         data: dataReducer,
         errors: errorsReducer,
         statuses: statusesReducer,
@@ -22,7 +24,7 @@ export default function countryReducer(state = {}, action) {
  * @param {Action} action
  * @return {?Country[]}
  */
-function dataReducer(data, action) {
+function dataReducer(data: any[], action: Action): any[] {
     switch (action.type) {
     case actionTypes.LOAD_COUNTRIES_SUCCEEDED:
         return action.payload || [];
@@ -38,7 +40,7 @@ function dataReducer(data, action) {
  * @param {Action} action
  * @return {Object}
  */
-function errorsReducer(errors = {}, action) {
+function errorsReducer(errors: any = {}, action: Action): any {
     switch (action.type) {
     case actionTypes.LOAD_COUNTRIES_REQUESTED:
     case actionTypes.LOAD_COUNTRIES_SUCCEEDED:
@@ -58,7 +60,7 @@ function errorsReducer(errors = {}, action) {
  * @param {Action} action
  * @return {Object}
  */
-function statusesReducer(statuses = {}, action) {
+function statusesReducer(statuses: any = {}, action: Action): any {
     switch (action.type) {
     case actionTypes.LOAD_COUNTRIES_REQUESTED:
         return { ...statuses, isLoading: true };
